Separate team sections; use old fashioned sticky detect
