+++ package.json
@@ -31,6 +31,7 @@
 		"firebase": "^10.12.4",
 		"firebase-admin": "^12.2.0",
 		"framer-motion": "^11.3.8",
+		"lodash-es": "^4.17.21",
 		"lucide-react": "^0.411.0",
 		"next": "14.2.5",
 		"next-themes": "^0.3.0",
