+++ src/app/teams/[index]/page.js
@@ -5,14 +5,9 @@ import useSWR from "swr";
 import fetcher from "@/lib/fetcher";
 
 // Components & UI
-import Anchor from "@/components/main/common/anchor";
 import TeamHeader from "@/components/main/team/team-header";
-import { MarkdownParagraph } from "@/components/main/common/paragraph";
+import { DescSection, IntroSection } from "@/components/main/team/team-sections";
 import { Skeleton } from "@/components/ui/skeleton";
-
-// Icon & Images
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faEnvelope } from "@fortawesome/free-regular-svg-icons";
 
 
 
@@ -39,59 +34,11 @@ export default function TeamPage({ params }) {
 			<main>
 				<div className="wrapper-md">
 					<article className="grid gap-y-18 py-12">
-						<IntroductionSection introduction={teamData.introduction} />
-						<TeamDescSection teamDesc={teamData.team_desc} members={teamData.members} />
+						<IntroSection introduction={teamData.introduction} />
+						<DescSection teamDesc={teamData.team_desc} members={teamData.members} />
 					</article>
 				</div>
             </main>
         </>
-	);
-}
-
-function IntroductionSection({ className, introduction }) {
-	return (
-		<TeamSection className={`${className}`}>
-			<h2 className="whitespace-nowrap">作品介紹</h2>
-			<div>
-				<MarkdownParagraph content={introduction} />
-			</div>
-		</TeamSection>
-	);
-}
-
-function TeamDescSection({ className, teamDesc, members = [] }) {
-	return (
-		<TeamSection className={`${className}`}>
-			<h2 className="whitespace-nowrap">隊伍介紹</h2>
-			<div>
-				<MarkdownParagraph content={teamDesc} />
-				<h3 className="mb-2">成員</h3>
-				<div className="grid gap-4">
-					{members.map((member, index) => (
-						<div key={index}>
-							<h4>
-								{member.name}
-								{member?.email && (
-									<Anchor href={`mailto:${member.email}`} className="ml-2">
-										<FontAwesomeIcon icon={faEnvelope} />
-									</Anchor>
-								)}
-							</h4>
-							{member?.thoughts && (
-								<MarkdownParagraph content={member.thoughts} />
-							)}
-						</div>
-					))}
-				</div>
-			</div>
-		</TeamSection>
-	);
-}
-
-function TeamSection({ className, children }) {
-	return (
-		<section className={`grid gap-3 md:grid-cols-[10rem,_1fr] lg:grid-cols-[16rem,_1fr] xl:grid-cols-[24rem,_1fr] ${className}`}>
-			{children}
-		</section>
 	);
 }
